- Fixing rename bug 1/2.
