+++ src/constants/CourseURLSuffixTEST.ts
